Add form submission and feedback routing functionality
Implemented submit functionality on the "Contact Us" form and clicking "Submit" now saves the filled-in form data. Created a new SavePage that displays the received feedback from the form and includes an auto-rerouting feature back to the homepage. Also introduced a new ContactUsFormModel to handle form data. Form state and input data sharing is handled by a new DataService.